EnchantmentUtils - change up how classinfo is done
